feat: 백준 문제 생성 오류 해결 (#20)

+++ app/src/main/java/com/example/softwareproject/data/dto/problem/BaekjoonProblemDto.kt
@@ -6,7 +6,7 @@ data class BaekjoonProblemDto (
     val problemId: String,
     val titleKo: String,
     val acceptedUserCount: Int,
-    val averageTries: Int,
+    val averageTries: Double,
     val tags: List<TagDto>
 
 )
